requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
coloredlogs==15.0.1
exceptiongroup==1.2.2
fastapi==0.115.8
flatbuffers==25.1.24
h11==0.14.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lazy_loader==0.4
llvmlite==0.43.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
opencv-python-headless==4.11.0.86
packaging==24.2
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
protobuf==5.29.3
pydantic==2.10.6
pydantic_core==2.27.2
PyMatting==1.1.13
python-multipart==0.0.20
referencing==0.36.2
rembg==2.0.61
requests==2.32.3
rpds-py==0.22.3
scikit-image==0.24.0
scipy==1.13.1
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.3
tifffile==2024.8.30
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
